+++ HISTORY.rst
@@ -6,8 +6,10 @@ History
 Current
 +++++++
 
+* Better infrastructure (type annotations, documentation, ...)
 * Dropped support for Python 2.
 * New function: update_observation()
+* rest_api.AuthenticationError is now exceptions.AuthenticationError
 
 
 0.1.0 (2018-10-10)